requirements.txt
asgiref==3.8.1
Django==4.2
django-filter==24.2
djangorestframework==3.15.1
mutagen==1.47.0
pillow==10.3.0
sqlparse==0.5.0
typing_extensions==4.11.0
